[Refactor][2/N] Move out orc input stream and min max decoder.
