+++ be/src/formats/CMakeLists.txt
@@ -22,7 +22,9 @@ add_library(Formats STATIC
         json/numeric_column.cpp
         json/binary_column.cpp
         orc/orc_chunk_reader.cpp
+        orc/orc_input_stream.cpp
         orc/orc_mapping.cpp
+        orc/orc_min_max_decoder.cpp
         orc/fill_function.cpp
         orc/utils.cpp
         parquet/column_chunk_reader.cpp
