+++ be/src/formats/orc/orc_chunk_reader.cpp
@@ -17,6 +17,7 @@
 #include "exprs/vectorized/cast_expr.h"
 #include "exprs/vectorized/literal.h"
 #include "formats/orc/fill_function.h"
+#include "formats/orc/orc_input_stream.h"
 #include "formats/orc/orc_mapping.h"
 #include "fs/fs.h"
 #include "gen_cpp/orc_proto.pb.h"
@@ -1126,172 +1127,6 @@ void OrcChunkReader::set_conjuncts_and_runtime_filters(const std::vector<Expr*>&
     }
 }
 
-#define DOWN_CAST_ASSIGN_MIN_MAX(TYPE)                         \
-    do {                                                       \
-        ColumnHelper::cast_to_raw<TYPE>(min_col)->append(min); \
-        ColumnHelper::cast_to_raw<TYPE>(max_col)->append(max); \
-        return Status::OK();                                   \
-    } while (0)
-
-static Status decode_int_min_max(PrimitiveType ptype, const orc::proto::ColumnStatistics& colStats,
-                                 const ColumnPtr& min_col, const ColumnPtr& max_col) {
-    if (colStats.has_intstatistics() && colStats.intstatistics().has_minimum() &&
-        colStats.intstatistics().has_maximum()) {
-        const auto& stats = colStats.intstatistics();
-        int64_t min = stats.minimum();
-        int64_t max = stats.maximum();
-
-        switch (ptype) {
-        case PrimitiveType::TYPE_TINYINT:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_TINYINT);
-        case PrimitiveType::TYPE_SMALLINT:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_SMALLINT);
-        case PrimitiveType::TYPE_INT:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_INT);
-        case PrimitiveType::TYPE_BIGINT:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_BIGINT);
-        default:
-            break;
-        }
-    }
-    return Status::NotFound("int column stats not found");
-}
-
-static Status decode_double_min_max(PrimitiveType ptype, const orc::proto::ColumnStatistics& colStats,
-                                    const ColumnPtr& min_col, const ColumnPtr& max_col) {
-    if (colStats.has_doublestatistics() && colStats.doublestatistics().has_minimum() &&
-        colStats.doublestatistics().has_maximum()) {
-        const auto& stats = colStats.doublestatistics();
-        double min = stats.minimum();
-        double max = stats.maximum();
-        switch (ptype) {
-        case PrimitiveType::TYPE_FLOAT:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_FLOAT);
-        case PrimitiveType::TYPE_DOUBLE:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_DOUBLE);
-        default:
-            break;
-        }
-    }
-    return Status::NotFound("double column stats not found");
-}
-static Status decode_string_min_max(PrimitiveType ptype, const orc::proto::ColumnStatistics& colStats,
-                                    const ColumnPtr& min_col, const ColumnPtr& max_col) {
-    if (colStats.has_stringstatistics() && colStats.stringstatistics().has_minimum() &&
-        colStats.stringstatistics().has_maximum()) {
-        const auto& stats = colStats.stringstatistics();
-        const std::string& min_value = stats.minimum();
-        const std::string& max_value = stats.maximum();
-        size_t min_value_size = min_value.size();
-        size_t max_value_size = max_value.size();
-        if (ptype == TYPE_CHAR) {
-            min_value_size = remove_trailing_spaces(min_value.c_str(), min_value_size);
-            max_value_size = remove_trailing_spaces(max_value.c_str(), max_value_size);
-        }
-        const Slice min(min_value.c_str(), min_value_size);
-        const Slice max(max_value.c_str(), max_value_size);
-        switch (ptype) {
-        case PrimitiveType::TYPE_VARCHAR:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_VARCHAR);
-        case PrimitiveType::TYPE_CHAR:
-            DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_CHAR);
-        default:
-            break;
-        }
-    }
-    return Status::NotFound("string column stats not found");
-}
-
-static Status decode_date_min_max(PrimitiveType ptype, const orc::proto::ColumnStatistics& colStats,
-                                  const ColumnPtr& min_col, const ColumnPtr& max_col) {
-    if (colStats.has_datestatistics() && colStats.datestatistics().has_minimum() &&
-        colStats.datestatistics().has_maximum()) {
-        const auto& stats = colStats.datestatistics();
-        DateValue min, max;
-        OrcDateHelper::orc_date_to_native_date(&min, stats.minimum());
-        OrcDateHelper::orc_date_to_native_date(&max, stats.maximum());
-        DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_DATE);
-    }
-    return Status::NotFound("date column stats not found");
-}
-
-// It's quite odd that, timestamp statistics stores milliseconds since unix epoch time.
-// but timestamp column vector batch stores seconds since unix epoch time.
-// https://orc.apache.org/specification/ORCv1/
-static Status decode_datetime_min_max(PrimitiveType ptype, const orc::proto::ColumnStatistics& colStats,
-                                      int64_t tz_offset_in_seconds, const ColumnPtr& min_col,
-                                      const ColumnPtr& max_col) {
-    if (colStats.has_timestampstatistics() && colStats.timestampstatistics().has_minimumutc() &&
-        colStats.timestampstatistics().has_maximumutc()) {
-        const auto& stats = colStats.timestampstatistics();
-        TimestampValue min, max;
-        const cctz::time_zone utc_tzinfo = cctz::utc_time_zone();
-        {
-            int64_t ms = stats.minimumutc();
-            int64_t ns = 0;
-            if (stats.has_minimumnanos()) {
-                ns = stats.minimumnanos();
-            }
-            int64_t secs = ms / 1000;
-            ns += (ms - secs * 1000) * 1000000L;
-            OrcTimestampHelper::orc_ts_to_native_ts(&min, utc_tzinfo, tz_offset_in_seconds, secs, ns);
-        }
-
-        {
-            int64_t ms = stats.maximumutc();
-            int64_t ns = 0;
-            if (stats.has_maximumnanos()) {
-                ns = stats.maximumnanos();
-            }
-            int64_t secs = ms / 1000;
-            ns += (ms - secs * 1000) * 1000000L;
-            OrcTimestampHelper::orc_ts_to_native_ts(&max, utc_tzinfo, tz_offset_in_seconds, secs, ns);
-        }
-
-        DOWN_CAST_ASSIGN_MIN_MAX(PrimitiveType::TYPE_DATETIME);
-    }
-    return Status::NotFound("date column stats not found");
-}
-
-Status OrcChunkReader::decode_min_max_value(SlotDescriptor* slot, const orc::proto::ColumnStatistics& stats,
-                                            ColumnPtr min_col, ColumnPtr max_col, int64_t tz_offset_in_seconds) {
-    if (slot->is_nullable()) {
-        auto* a = ColumnHelper::as_raw_column<NullableColumn>(min_col);
-        auto* b = ColumnHelper::as_raw_column<NullableColumn>(max_col);
-        a->mutable_null_column()->append(0);
-        b->mutable_null_column()->append(0);
-        min_col = a->data_column();
-        max_col = b->data_column();
-    }
-    PrimitiveType ptype = slot->type().type;
-    switch (ptype) {
-    case PrimitiveType::TYPE_TINYINT:
-    case PrimitiveType::TYPE_SMALLINT:
-    case PrimitiveType::TYPE_INT:
-    case PrimitiveType::TYPE_BIGINT:
-        // case PrimitiveType::TYPE_LARGEINT:
-        return decode_int_min_max(ptype, stats, min_col, max_col);
-
-    case PrimitiveType::TYPE_FLOAT:
-    case PrimitiveType::TYPE_DOUBLE:
-        return decode_double_min_max(ptype, stats, min_col, max_col);
-
-    case PrimitiveType::TYPE_VARCHAR:
-    case PrimitiveType::TYPE_CHAR:
-        return decode_string_min_max(ptype, stats, min_col, max_col);
-
-    case PrimitiveType::TYPE_DATE:
-        return decode_date_min_max(ptype, stats, min_col, max_col);
-
-    case PrimitiveType::TYPE_DATETIME:
-        return decode_datetime_min_max(ptype, stats, tz_offset_in_seconds, min_col, max_col);
-
-    default:
-        return Status::NotSupported("Not support to decode min/max from orc column stats. type = " +
-                                    std::to_string(ptype));
-    }
-}
-
 Status OrcChunkReader::apply_dict_filter_eval_cache(const std::unordered_map<SlotId, FilterPtr>& dict_filter_eval_cache,
                                                     Filter* filter) {
     if (dict_filter_eval_cache.size() == 0) {
@@ -1366,98 +1201,11 @@ int OrcChunkReader::get_column_id_by_slot_name(const std::string& slot_name) con
     return -1;
 }
 
-// ======================================================================================
-
-ORCHdfsFileStream::ORCHdfsFileStream(RandomAccessFile* file, uint64_t length)
-        : _file(file), _length(length), _cache_buffer(0), _cache_offset(0), _buffer_stream(_file) {
-    SharedBufferedInputStream::CoalesceOptions options = {.max_dist_size = config::io_coalesce_read_max_distance_size,
-                                                          .max_buffer_size = config::io_coalesce_read_max_buffer_size};
-    _buffer_stream.set_coalesce_options(options);
-}
-
-void ORCHdfsFileStream::prepareCache(orc::InputStream::PrepareCacheScope scope, uint64_t offset, uint64_t length) {
-    const size_t cache_max_size = config::orc_file_cache_max_size;
-    if (length > cache_max_size) return;
-    if (canUseCacheBuffer(offset, length)) return;
-
-    // If this stripe is small, probably other stripes are also small
-    // we combine those reads into one, and try to read several stripes in one shot.
-    if (scope == orc::InputStream::PrepareCacheScope::READ_FULL_STRIPE) {
-        length = std::min(_length - offset, cache_max_size);
-    }
-
-    _cache_buffer.resize(length);
-    _cache_offset = offset;
-    doRead(_cache_buffer.data(), length, offset, true);
-}
-
-bool ORCHdfsFileStream::canUseCacheBuffer(uint64_t offset, uint64_t length) {
-    if ((_cache_buffer.size() != 0) && (offset >= _cache_offset) &&
-        ((offset + length) <= (_cache_offset + _cache_buffer.size()))) {
-        return true;
-    }
-    return false;
-}
-
-void ORCHdfsFileStream::read(void* buf, uint64_t length, uint64_t offset) {
-    if (canUseCacheBuffer(offset, length)) {
-        size_t idx = offset - _cache_offset;
-        memcpy(buf, _cache_buffer.data() + idx, length);
-    } else {
-        doRead(buf, length, offset, false);
-    }
-}
-
-const std::string& ORCHdfsFileStream::getName() const {
-    return _file->filename();
-}
-
-void ORCHdfsFileStream::doRead(void* buf, uint64_t length, uint64_t offset, bool direct) {
-    if (buf == nullptr) {
-        throw orc::ParseError("Buffer is null");
-    }
-    Status status;
-    if (direct || !_buffer_stream_enabled) {
-        status = _file->read_at_fully(offset, buf, length);
-    } else {
-        const uint8_t* ptr = nullptr;
-        size_t nbytes = length;
-        status = _buffer_stream.get_bytes(&ptr, offset, &nbytes, false);
-        DCHECK_EQ(nbytes, length);
-        if (status.ok()) {
-            ::memcpy(buf, ptr, length);
-        }
-    }
-    if (!status.ok()) {
-        auto msg = strings::Substitute("Failed to read $0: $1", _file->filename(), status.to_string());
-        throw orc::ParseError(msg);
-    }
-}
-
-void ORCHdfsFileStream::clearIORanges() {
-    _buffer_stream_enabled = false;
-    _buffer_stream.release();
-}
-
-void ORCHdfsFileStream::setIORanges(std::vector<orc::InputStream::IORange>& io_ranges) {
-    _buffer_stream_enabled = true;
-    std::vector<SharedBufferedInputStream::IORange> bs_io_ranges;
-    bs_io_ranges.reserve(io_ranges.size());
-    for (const auto& r : io_ranges) {
-        bs_io_ranges.emplace_back(SharedBufferedInputStream::IORange{.offset = static_cast<int64_t>(r.offset),
-                                                                     .size = static_cast<int64_t>(r.size)});
-    }
-    Status st = _buffer_stream.set_io_ranges(bs_io_ranges);
-    if (!st.ok()) {
-        auto msg = strings::Substitute("Failed to setIORanges $0: $1", _file->filename(), st.to_string());
-        throw orc::ParseError(msg);
-    }
-}
-
 bool OrcChunkReader::is_implicit_castable(TypeDescriptor& starrocks_type, const TypeDescriptor& orc_type) {
     if (starrocks_type.is_decimal_type() && orc_type.is_decimal_type()) {
         return true;
     }
     return false;
 }
+
 } // namespace starrocks::vectorized
